It's unclear which should win if a flyweight is created with readonly differing between instances
So I'll throw an error about it

Alternately I could set it to non-readonly and give a warning, as readonly is only there to state your intention and you're probably not going to try modifying it in the code path you create it.
On the other hand, if you're creating it not readonly, you probably need to modify it.